fix: include dagrun conf (#124)
* fix: don't ignore dagrun conf

The dagrun conf is stored as JSONB in the database by pickling the
corresponding object. Unlike the executor config, the dagrun conf is
always a JSON-serializable dictionary and thus we won't run into
problems of JSON serializing foreign types.

* ci: ensure tests are run with a compatible airflow version

* bump version to 2.2.4

* fix: return conf in unpickled form

Neither do we want to return the conf in pickled binary form, nor do
tests.

* test: fix manual tests

* use `delete` instead of `post` for deletion
* separate variables for dagruns and taskinstances
* delete dagrun after task instance tests