add Airflow 3 support (#147)
* refactor: rename compat module to af2

This will be the place for all Airflow 2 specific code.

* refactor: move starship plugins to af2 module

These are being made accessible to Airflow's plugin entrypoints via astronomer_starship.plugins.

* refactor: make sure Airflow 2 specific code only loads on Airflow 2

* refactor: make the `compat` module the access point for imports

* refactor: move version specific code into private module

Any imports should happen via the compat module.

* fix: add StarshipApi to compat module

* bump: set version to 2.7.0

* fix: turn `starship_compatability` into a module again

* fix: use `Tuple` from `typing` for compatibility

* refactor: revert to original imports for entrypoints and API docs

* refactor: move telescope code to common module

* refactor: replace use of jsonify with HttpError

We jsonify the corresponding error inside the flask specific starship_route function.

* refactor: replace use of no content response with None result

We turn the None result into a no content response in `starship_route`.

* refactor: move flask.Response import to function scope

* test: adjust assertions for NO_CONTENT responses'

* test: set test paths

* refactor: move generic functionality into common module

* test: remove session scope for compatibility layer fixture

The instance's DB session shall not be shared across requests.

* ci: use Ruff for linting and formatting

We replace the following tools with Ruff and extend the selected rules:

* black
* blacken-docs
* isort

* ci: move constraints into separate file

We want to keep flask-limiter contrained, but we don't want to install
it if it's not needed (i.e on Airflow 3.x).

* chore: add classifiers to package metadata

* refactor: move common parts of StarshipAirflow to common base class

This base class will be used in Airflow 2 and Airflow 3 compatibility layers
and is Airflow version agnostic.

* refactor: move functionality for pools, variables and connections to BaseStarshipAirflow

* refactor: rename dev to dev2

The dev2 folder is for Airflow 2 development, while dev3 will be for Airflow 3 development.

* dev: add dev environment for Airflow 3

* ci: replace pre-commit with prek in GitHub Actions

It's faster and we don't need to setup python or just.

* ci: update pre-commit hooks

* feat: add initial Airflow 3 support for the Starship API

The following API endpoints are implemented:

* health
* telescope
* airflow version
* info
* env_vars
* pools
* variables
* connections

* refactor: add missing endpoints in Airflow 3's StarshipApi

All needed methods in BaseStarshipAirflow are defined and have to be
either replaced by a common implementation or be implemented in the
corresponding subclass.

* docs: move API docs to api.md

We know have two different StarshipApi implementations and can't easily
maintain doc strings in either. By moving into the markdown file it's
also easier to adhere to certain markdown formatting rules.

* docs: use call-outs for deprecation and experimental hints

* test: adapt dagbag print command for Airflow 3

In Airflow 3 all dags are stored serialized, so the argument
`store_serialized_dags` has been removed.

* DAG and DAG Runs

* dev: Task Instances Support and Dag Version

* test: add test support

* dev: AF3 starship UI (#148)

* dev: AF3 starship UI

* test: add tests and codecov

* dev: UI Fixes

* dev: add compression to api calls

* refactor: remove endpoint to update dag version IDs

For now we don't support migrating DAG bundles. The existing code
is able to operate without this, while setting the dag version ID for
task instance and dag runs to the latest version automatically.

In future version we can consider adding full supported and documented
API endpoints to migrate DAG bundles.

* dev: build frontend for Airflow 3 dev instance

Since we added the UI to Airflow 3 too, we also need to build the frontend
when we start or reload the dev3 instance.

* chore: update package-lock.json

* refactor: don't recreate TI UUID's

This will make it easier to migrate dependent models in the future.

We should not be worrying about collisions, since the original UUIDs
are already pretty random.

---------

Co-authored-by: Akshay <akshaykumar.salunke@astronomer.io>
Co-authored-by: akshaykumarsalunke <165752380+akshaykumarsalunke@users.noreply.github.com>